feature/list_filter, * fix admin page components styles, * fix action file name * fix users reducer name * admin page components, remove styles from components
