+++ src/actions/index.js
@@ -1,3 +1,10 @@
 export { fetchUser, fetchUserSuccess, saveProfile } from './user';
+export {
+  getAllUsers,
+  getUserData,
+  reciveAllUser,
+  setSelectedUser,
+  setFilter,
+} from './users.actions';
 
 export { serverError } from './errors';
