+++ src/components/Admin/UserList/style.js
@@ -1,0 +1,30 @@
+const styles = () => ({
+  root: {
+    width: '100%',
+    maxWidth: 320,
+    maxHeight: '70%',
+  },
+  searchWrapper: {
+    display: 'flex',
+    paddingLeft: '20px',
+    alighnItems: 'center',
+    margin: '0px 3px 5px 8px',
+    marginBottom: '20px',
+  },
+  listWrapper: {
+    overflowY: 'scroll',
+    maxHeight: '300',
+    margin: '0px 3px 5px 8px',
+  },
+  input: {
+    marginLeft: 8,
+    flex: 1,
+  },
+  divider: {
+    width: 1,
+    height: 36,
+    margin: 4,
+  },
+});
+
+export default styles;
