+++ src/components/Admin/UserListItem/style.js
@@ -1,0 +1,7 @@
+const styles = () => ({
+  itemListPadding: {
+    padding: '2px 16px',
+  },
+});
+
+export default styles;
