+++ src/pages/Admin/style.js
@@ -1,0 +1,15 @@
+const styles = () => ({
+  wrapper: {
+    display: 'flex',
+    maxHeight: '300px',
+    paddingTop: '20px',
+  },
+  infoHeader: {
+    width: '100%',
+    textAlign: 'center',
+    color: 'white',
+    height: 30,
+  },
+});
+
+export default styles;
