+++ src/sagas/adminSaga/index.js
@@ -1,5 +1,5 @@
 import { take, put, call, fork, select, takeEvery, all } from 'redux-saga/effects'; //eslint-disable-line
-import * as actions from 'actions/adminPage';
+import * as actions from 'actions';
 import { GET_ALL_USERS } from 'actions/types';
 import fb from 'api/firebase/index';
